@@ -5,30 +5,29 @@ using FIsionAPI.Data.Contexto;
 using FIsionAPI.Data.Repositorios;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace FIsionAPI.API.Configurations
+namespace FIsionAPI.API.Configurations;
+
+public static class DepedenceInjectionConfig
 {
-    public static class DepedenceInjectionConfig
+    public static IServiceCollection ResolveDependecies(this IServiceCollection service)
     {
-        public static IServiceCollection ResolveDependecies(this IServiceCollection service)
-        {
-            service.AddScoped<FisionContext>();
-            service.AddScoped<IPessoaRepositoy, PessoaRepository>();
-            service.AddScoped<IEntidadeRepository, EntidadeRepository>();
-            service.AddScoped<EspecialidadeRepository, EspecialidadeRepository>();
-            service.AddScoped<IEnderecoPessoaRepository, EnderecoPessoaRepository>();
-            service.AddScoped<IContratoFinanceiroRepository, ContratoFinanceiroRepository>();
-            service.AddScoped<IMovimentoFinanceiroRepository, MovimentoFinanceiroRepository>();
-            service.AddScoped<ICaixaRepository, CaixaRepository>();
-            service.AddScoped<ICaixaService, CaixaService>();
-            service.AddScoped<IMovimentoFinanceiroAvulsoRepository, MovimentoFinanceiroAvulsoRepository>();
+        service.AddScoped<FisionContext>();
+        service.AddScoped<IPessoaRepositoy, PessoaRepository>();
+        service.AddScoped<IEntidadeRepository, EntidadeRepository>();
+        service.AddScoped<EspecialidadeRepository, EspecialidadeRepository>();
+        service.AddScoped<IEnderecoPessoaRepository, EnderecoPessoaRepository>();
+        service.AddScoped<IContratoFinanceiroRepository, ContratoFinanceiroRepository>();
+        service.AddScoped<IMovimentoFinanceiroRepository, MovimentoFinanceiroRepository>();
+        service.AddScoped<ICaixaRepository, CaixaRepository>();
+        service.AddScoped<ICaixaService, CaixaService>();
+        service.AddScoped<IMovimentoFinanceiroAvulsoRepository, MovimentoFinanceiroAvulsoRepository>();
 
-            service.AddScoped<IEntidadeService, EntidadeService>();
-            service.AddScoped<IMovimentoFinanceiroService, MovimentoFinanceiroService>();
-            service.AddScoped<IMovimentoFinanceiroAvulsoService, MovimentoFinanceiroAvulsoService>();
+        service.AddScoped<IEntidadeService, EntidadeService>();
+        service.AddScoped<IMovimentoFinanceiroService, MovimentoFinanceiroService>();
+        service.AddScoped<IMovimentoFinanceiroAvulsoService, MovimentoFinanceiroAvulsoService>();
 
-            service.AddScoped<INotificador, Notificador>();
+        service.AddScoped<INotificador, Notificador>();
 
-            return service;
-        }
+        return service;
     }
 }
