@@ -2,60 +2,66 @@
 using FIsionAPI.Business.Notificacões;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.ModelBinding;
-using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Threading.Tasks;
 
-namespace FIsionAPI.API.Controllers
+namespace FIsionAPI.API.Controllers;
+
+[ApiController]
+public class BaseController : ControllerBase
 {
-    [ApiController]
-    public class BaseController : ControllerBase
+    private readonly INotificador _notificador;
+
+    public BaseController(INotificador notificador)
     {
-        private readonly INotificador _notificador;
+        _notificador = notificador;
+    }
 
-        public BaseController(INotificador notificador)
-        {
-            _notificador = notificador;
-        }
-        protected bool OperacaoValida()
-        {
-            return !_notificador.TemNotificacao();
-        }
-        protected ActionResult CustomResponse(object result = null)
-        {
-            if (OperacaoValida())
-            {
-                return Ok(new
-                {
-                    success = true,
-                    data = result
-                });
-            }
+    protected bool OperacaoValida()
+    {
+        return !_notificador.TemNotificacao();
+    }
 
-            return BadRequest(new
+    protected ActionResult CustomResponse(object result = null)
+    {
+        if (OperacaoValida())
+        {
+            return Ok(new
             {
-                success = false,
-                erros = _notificador.ObterNotificacoes().Select(n=>n.Mensagem)
+                success = true,
+                data = result
             });
         }
-        protected ActionResult CustomReponse(ModelStateDictionary modelState)
+
+        return BadRequest(new
         {
-            if (!modelState.IsValid) NotificarErroModelInvalida(modelState);
-            return CustomResponse();
-        }
-        protected void NotificarErroModelInvalida(ModelStateDictionary modelState)
+            success = false,
+            erros = _notificador.ObterNotificacoes().Select(n => n.Mensagem)
+        });
+    }
+
+    protected ActionResult CustomReponse(ModelStateDictionary modelState)
+    {
+        if (!modelState.IsValid)
         {
-            var erros = modelState.Values.SelectMany(e => e.Errors);
-            foreach (var erro in erros)
-            {
-                var errorMsg = erro.Exception == null ? erro.ErrorMessage : erro.Exception.Message;
-                NotificarErro(errorMsg);
-            }
+            NotificarErroModelInvalida(modelState);
         }
-        protected void NotificarErro(string mensagem)
+
+        return CustomResponse();
+    }
+
+    protected void NotificarErroModelInvalida(ModelStateDictionary modelState)
+    {
+        var erros = modelState.Values.SelectMany(e => e.Errors);
+
+        foreach (var erro in erros)
         {
-            _notificador.Handle(new Notificacao(mensagem));
+            var errorMsg = erro.Exception == null ? erro.ErrorMessage : erro.Exception.Message;
+            NotificarErro(errorMsg);
         }
+    }
+
+    protected void NotificarErro(string mensagem)
+    {
+        _notificador.Handle(new Notificacao(mensagem));
     }
 }
