@@ -1,23 +1,20 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
 
-namespace FIsionAPI.API.ViewModels
+namespace FIsionAPI.API.ViewModels;
+
+public class CaixaViewModel
 {
-    public class CaixaViewModel
-    {
+    [Key]
+    public Guid Id { get; set; }
 
-        [Key]
-        public Guid Id { get; set; }
-        [Required]
-        public string Competencia { get; set; }
-        public decimal ValorDespesa { get; set; }
-        public decimal ValorReceita { get; set; }
-        public int Situacao { get; set; }
-        public string Status { get; set; }
-        public DateTime DataCadastro { get; set; }
-        public IEnumerable<MovimentoFinanceiroViewModel> Movimentos { get; set; }
-    }
+    [Required]
+    public string Competencia { get; set; }
+    public decimal ValorDespesa { get; set; }
+    public decimal ValorReceita { get; set; }
+    public int Situacao { get; set; }
+    public string Status { get; set; }
+    public DateTime DataCadastro { get; set; }
+    public IEnumerable<MovimentoFinanceiroViewModel> Movimentos { get; set; }
 }
