@@ -1,24 +1,21 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
 
-namespace FIsionAPI.API.ViewModels
+namespace FIsionAPI.API.ViewModels;
+
+public class EntidadeViewModel
 {
-    public class EntidadeViewModel
-    {
-        [Key]
-        public Guid Id { get; set; }
-        public DateTime DtEntrada { get; set; }
-        public DateTime DtSaida { get; set; }
-        public string Matricula { get; set; }
-        public PessoaViewModel Pessoa { get; set; }
-        public Guid PessoaId { get; set; }
-        public ContratoFinanceiroViewModel Contrato { get; set; }
-        public DateTime DataCadastro { get; set; }
-        public int Classe { get; set; }
-        public Guid? EspecialidadeId { get; set; }
-        public IEnumerable<EspecialidadeViewModel>? Especialidades { get; set; }
-    }
+    [Key]
+    public Guid Id { get; set; }
+    public DateTime DtEntrada { get; set; }
+    public DateTime DtSaida { get; set; }
+    public string Matricula { get; set; }
+    public PessoaViewModel Pessoa { get; set; }
+    public Guid PessoaId { get; set; }
+    public ContratoFinanceiroViewModel Contrato { get; set; }
+    public DateTime DataCadastro { get; set; }
+    public int Classe { get; set; }
+    public Guid? EspecialidadeId { get; set; }
+    public IEnumerable<EspecialidadeViewModel>? Especialidades { get; set; }
 }
