@@ -2,15 +2,13 @@
 using FIsionAPI.Business.Models.Enums;
 using System;
 using System.Collections.Generic;
-using System.Text;
 using System.Threading.Tasks;
 
-namespace FIsionAPI.Business.Interfaces
+namespace FIsionAPI.Business.Interfaces;
+
+public interface IContratoFinanceiroRepository : IRepository<ContratoFinanceiro>, IDisposable
 {
-    public interface IContratoFinanceiroRepository : IRepository<ContratoFinanceiro>, IDisposable
-    {
-        Task<ContratoFinanceiro> ObterContratoPorId(Guid id);
-        Task<IEnumerable<ContratoFinanceiro>> ObterContratoPorTipo(TipoContrato CodTipo);
-        Task<ContratoFinanceiro> ObterContratoEntidade(Guid EntidadeId);
-    }
+    Task<ContratoFinanceiro> ObterContratoPorId(Guid id);
+    Task<IEnumerable<ContratoFinanceiro>> ObterContratoPorTipo(TipoContrato CodTipo);
+    Task<ContratoFinanceiro> ObterContratoEntidade(Guid EntidadeId);
 }
