@@ -3,11 +3,10 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace FIsionAPI.Business.Interfaces
+namespace FIsionAPI.Business.Interfaces;
+
+public interface IEspecialidadeRepository : IRepository<Especialidades>, IDisposable
 {
-    public interface IEspecialidadeRepository : IRepository<Especialidades>, IDisposable
-    {
-        Task<IEnumerable<Especialidades>> ObterEspecialidades();
-        Task<Especialidades> ObterEspecialidadePorId(Guid id);
-    }
+    Task<IEnumerable<Especialidades>> ObterEspecialidades();
+    Task<Especialidades> ObterEspecialidadePorId(Guid id);
 }
