@@ -1,14 +1,11 @@
 ﻿using FIsionAPI.Business.Models;
 using System;
-using System.Collections.Generic;
-using System.Text;
 using System.Threading.Tasks;
 
-namespace FIsionAPI.Business.Interfaces
+namespace FIsionAPI.Business.Interfaces;
+
+public interface IMovimentoFinanceiroAvulsoService : IDisposable
 {
-    public interface IMovimentoFinanceiroAvulsoService : IDisposable
-    {
-        Task<bool> AdicionarMovimentoFinanceiroAvulso(MovimentoFinanceiroAvulso movimento);
-        Task<bool> AtualizarMovimentoFinanceiroAvulso(MovimentoFinanceiroAvulso movimento);
-    }
+    Task<bool> AdicionarMovimentoFinanceiroAvulso(MovimentoFinanceiroAvulso movimento);
+    Task<bool> AtualizarMovimentoFinanceiroAvulso(MovimentoFinanceiroAvulso movimento);
 }
