@@ -2,21 +2,17 @@
 using FIsionAPI.Business.Models.Enums;
 using System;
 using System.Collections.Generic;
-using System.Text;
 using System.Threading.Tasks;
 
-namespace FIsionAPI.Business.Interfaces
+namespace FIsionAPI.Business.Interfaces;
+
+public interface IMovimentoFinanceiroRepository : IRepository<MovimentoFinanceiroEntidade>, IDisposable
 {
-    public interface IMovimentoFinanceiroRepository : IRepository<MovimentoFinanceiroEntidade>, IDisposable
-    {
-        Task<MovimentoFinanceiroEntidade> BuscarMovimentoPorId(Guid id);
-        Task<IEnumerable<MovimentoFinanceiroEntidade>> BuscarMovimentos();
-        Task<IEnumerable<MovimentoFinanceiroEntidade>> BuscarMovimentosPorTipo(TipoMovimento CodTipo);
-        Task<IEnumerable<MovimentoFinanceiroEntidade>> BuscarMovimentoPorClasse(ClasseMovimento CodClasse);
-        Task<IEnumerable<MovimentoFinanceiroEntidade>> BuscarMovimentosPorContratoId(Guid ContratoId);
-        Task<MovimentoFinanceiroEntidade> BuscarMovimentosPorCompetencia(MovimentoFinanceiroEntidade movimento);
-        Task<IEnumerable<MovimentoFinanceiroEntidade>> ObterMovimentoPorTipoCompetencia(TipoMovimento tipo, string competencia);
-
-
-    }
+    Task<MovimentoFinanceiroEntidade> BuscarMovimentoPorId(Guid id);
+    Task<IEnumerable<MovimentoFinanceiroEntidade>> BuscarMovimentos();
+    Task<IEnumerable<MovimentoFinanceiroEntidade>> BuscarMovimentosPorTipo(TipoMovimento CodTipo);
+    Task<IEnumerable<MovimentoFinanceiroEntidade>> BuscarMovimentoPorClasse(ClasseMovimento CodClasse);
+    Task<IEnumerable<MovimentoFinanceiroEntidade>> BuscarMovimentosPorContratoId(Guid ContratoId);
+    Task<MovimentoFinanceiroEntidade> BuscarMovimentosPorCompetencia(MovimentoFinanceiroEntidade movimento);
+    Task<IEnumerable<MovimentoFinanceiroEntidade>> ObterMovimentoPorTipoCompetencia(TipoMovimento tipo, string competencia);
 }
