@@ -1,14 +1,11 @@
 ﻿using FIsionAPI.Business.Notificacões;
-using System;
 using System.Collections.Generic;
-using System.Text;
 
-namespace FIsionAPI.Business.Interfaces
+namespace FIsionAPI.Business.Interfaces;
+
+public interface INotificador
 {
-    public interface INotificador
-    {
-        bool TemNotificacao();
-        List<Notificacao> ObterNotificacoes();
-        void Handle(Notificacao notificacao);
-    }
+    bool TemNotificacao();
+    List<Notificacao> ObterNotificacoes();
+    void Handle(Notificacao notificacao);
 }
