@@ -1,15 +1,12 @@
 ﻿using FIsionAPI.Business.Models;
 using System;
-using System.Collections.Generic;
-using System.Text;
 using System.Threading.Tasks;
 
-namespace FIsionAPI.Business.Interfaces
+namespace FIsionAPI.Business.Interfaces;
+
+public interface IPessoaRepositoy : IRepository<Pessoa>
 {
-    public interface IPessoaRepositoy : IRepository<Pessoa>
-    {
-        Task<Pessoa> ObterPorPessoaId(Guid id);
-        Task<Pessoa> ObterPorCPF(string CPF);
-        Task<Pessoa> ObterEntidadePessoa(Guid id);
-    }
+    Task<Pessoa> ObterPorPessoaId(Guid id);
+    Task<Pessoa> ObterPorCPF(string CPF);
+    Task<Pessoa> ObterEntidadePessoa(Guid id);
 }
