@@ -2,18 +2,16 @@
 using System;
 using System.Collections.Generic;
 using System.Linq.Expressions;
-using System.Text;
 using System.Threading.Tasks;
 
-namespace FIsionAPI.Business.Interfaces
+namespace FIsionAPI.Business.Interfaces;
+
+public interface IRepository<TEntity> : IDisposable where TEntity : Entity
 {
-    public interface IRepository<TEntity> : IDisposable where TEntity : Entity
-    {
-        Task Adicionar(TEntity entity);
-        Task<TEntity> ObterPorId(Guid id);
-        Task<List<TEntity>> ObterTodos();
-        Task Atualizar(TEntity entity);
-        Task Remover(Guid Id);
-        Task<IEnumerable<TEntity>> Buscar(Expression<Func<TEntity, bool>> predicate);
-    }
+    Task Adicionar(TEntity entity);
+    Task<TEntity> ObterPorId(Guid id);
+    Task<List<TEntity>> ObterTodos();
+    Task Atualizar(TEntity entity);
+    Task Remover(Guid Id);
+    Task<IEnumerable<TEntity>> Buscar(Expression<Func<TEntity, bool>> predicate);
 }
