@@ -1,13 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Text;
 
-namespace FIsionAPI.Business.Models
+namespace FIsionAPI.Business.Models;
+
+public class Especialidades : Entity
 {
-    public class Especialidades : Entity
-    {
-        public string Descricao { get; set; }
-        public DateTime DataCadastro { get; set; }
-        public IEnumerable<Entidade> Entidades { get; set; }
-    }
+    public string Descricao { get; set; }
+    public DateTime DataCadastro { get; set; }
+    public IEnumerable<Entidade> Entidades { get; set; }
 }
