@@ -1,18 +1,17 @@
 ﻿using FIsionAPI.Business.Models.Enums;
 using System;
 
-namespace FIsionAPI.Business.Models
+namespace FIsionAPI.Business.Models;
+
+public class MovimentoFinanceiroAvulso : Entity
 {
-    public class MovimentoFinanceiroAvulso : Entity
-    {
-        public ClasseMovimento Classe { get; set; }
-        public decimal ValorTotal { get; set; }
-        public SituacaoMovimento Situacao { get; set; }
-        public DateTime DtPagamento { get; set; }
-        public string CompetenciaMensalidade { get; set; }
-        public string Descricao { get; set; }        
-        public DateTime DataCadastro { get; set; }
-        public Caixa Caixa { get; set; }
-        public TipoMovimento Tipo { get; set; }        
-    }
+    public ClasseMovimento Classe { get; set; }
+    public decimal ValorTotal { get; set; }
+    public SituacaoMovimento Situacao { get; set; }
+    public DateTime DataPagamento { get; set; }
+    public string CompetenciaMensalidade { get; set; }
+    public string Descricao { get; set; }
+    public DateTime DataCadastro { get; set; }
+    public Caixa Caixa { get; set; }
+    public TipoMovimento Tipo { get; set; }
 }
