@@ -2,30 +2,29 @@
 using System.Collections.Generic;
 using System.Linq;
 
-namespace FIsionAPI.Business.Notificacões
+namespace FIsionAPI.Business.Notificacões;
+
+public class Notificador : INotificador
 {
-    public class Notificador : INotificador
+    private List<Notificacao> _notificacoes;
+
+    public Notificador()
     {
-        private List<Notificacao> _notificacoes;
+        _notificacoes = new List<Notificacao>();
+    }
 
-        public Notificador()
-        {
-            _notificacoes = new List<Notificacao>();
-        }
+    public void Handle(Notificacao notificacao)
+    {
+        _notificacoes.Add(notificacao);
+    }
 
-        public void Handle(Notificacao notificacao)
-        {
-            _notificacoes.Add(notificacao);
-        }
+    public List<Notificacao> ObterNotificacoes()
+    {
+        return _notificacoes;
+    }
 
-        public List<Notificacao> ObterNotificacoes()
-        {
-            return _notificacoes;
-        }
-
-        public bool TemNotificacao()
-        {
-            return _notificacoes.Any();
-        }
+    public bool TemNotificacao()
+    {
+        return _notificacoes.Any();
     }
 }
