@@ -1,198 +1,238 @@
 ﻿using FIsionAPI.Business.Interfaces;
 using FIsionAPI.Business.Models;
+using FIsionAPI.Business.Models.Enums;
 using System;
 using System.Threading.Tasks;
 
-namespace FIsionAPI.Business.Services
+namespace FIsionAPI.Business.Services;
+
+public class MovimentoFinanceiroService : BaseService, IMovimentoFinanceiroService
 {
-    public class MovimentoFinanceiroService : BaseService, IMovimentoFinanceiroService
+    private readonly IMovimentoFinanceiroRepository _movimentoRepository;
+    private readonly IContratoFinanceiroRepository _contratoRepository;
+
+    public MovimentoFinanceiroService(INotificador notificador,
+                                      IContratoFinanceiroRepository contratoRepository,
+                                      IMovimentoFinanceiroRepository movimentoRepository) : base(notificador)
     {
-        private readonly IMovimentoFinanceiroRepository _movimentoRepository;
-        private readonly IContratoFinanceiroRepository _contratoRepository;        
-        public MovimentoFinanceiroService(INotificador notificador,                                            
-                                            IContratoFinanceiroRepository contratoRepository,                                            
-                                            IMovimentoFinanceiroRepository movimentoRepository)
-                                             : base(notificador)
+        _movimentoRepository = movimentoRepository;
+        _contratoRepository = contratoRepository;
+    }
+
+    public async Task<bool> AdicionarMensalidade(MovimentoFinanceiroEntidade movimento)
+    {
+        var contratoFinanceiro = await ObterContrato(movimento);
+
+        if (contratoFinanceiro == null)
         {
-            _movimentoRepository = movimentoRepository;
-            _contratoRepository = contratoRepository;            
+            Notificar("Contrato financeiro não encontrado");
+            return false;
         }
 
-        public async Task<bool> AdicionarMensalidade(MovimentoFinanceiroEntidade movimento)
+        if (EntidadeInativa(contratoFinanceiro.Entidade.DataSaida))
         {
-            
-            var contrato = await ObterContrato(movimento);                
-            
-
-            if (contrato.Entidade.DtSaida != DateTime.Parse("0001-01-01"))
-            {
-                Notificar("Entidade Inativa");
-                return false;
-            }
-
-            if (!await VerificarCompetencia(movimento))
-            {
-                Notificar("A mensalidade para a competência "+movimento.CompetenciaMensalidade+" já existe!");
-                return false;
-            }
-
-            MovimentoFinanceiroEntidade mensalidade = new MovimentoFinanceiroEntidade();
-
-            mensalidade.DataCadastro = DateTime.Now;
-            mensalidade.ValorMensal = contrato.ValorMensal;
-            mensalidade.Desconto = (contrato.ValorMensal * (contrato.Desconto) / 100);
-            mensalidade.ValorReceber = (contrato.ValorMensal) - (contrato.ValorMensal * (contrato.Desconto) / 100);
-            mensalidade.ValorPago = 0;
-            mensalidade.Classe = Models.Enums.ClasseMovimento.Aluno;
-            mensalidade.Situacao = Models.Enums.SituacaoMovimento.Aberto;
-            mensalidade.Tipo = Models.Enums.TipoMovimento.Receita;
-            mensalidade.ContratoId = contrato.Id;
-            mensalidade.CompetenciaMensalidade = movimento.CompetenciaMensalidade;
-            mensalidade.DtVencimento = DateTime.Parse(await DefinirDataVencimento(movimento));
-
-            await _movimentoRepository.Adicionar(mensalidade);
-            return true;
-
+            Notificar("Entidade Inativa");
+            return false;
         }
-        public async Task<bool> AdicionarValorProfissional(MovimentoFinanceiroEntidade movimento)
+
+        if (!await ExisteCompetencia(movimento))
         {
-            var contrato = await ObterContrato(movimento);
-
-            if (contrato.Entidade.DtSaida != DateTime.Parse("0001-01-01"))
-            {
-                Notificar("Entidade Inativa");
-                return false;
-            }
-
-            if (!await VerificarCompetencia(movimento))
-            {
-                Notificar("Esse Profissional já possui valor para a competência " + movimento.CompetenciaMensalidade);
-                return false;
-            }
-
-            MovimentoFinanceiroEntidade valorProfissional = new MovimentoFinanceiroEntidade();
-            valorProfissional.DataCadastro = DateTime.Now;
-            valorProfissional.ValorMensal = contrato.Quantidade * contrato.ValorUnitario;
-            valorProfissional.ValorPago = 0;            
-            valorProfissional.ValorReceber = ((contrato.ValorUnitario * contrato.Quantidade) * (contrato.MargemLucro) / 100);
-            valorProfissional.Situacao = Models.Enums.SituacaoMovimento.Aberto;
-            valorProfissional.Classe = Models.Enums.ClasseMovimento.Profissional;
-            valorProfissional.Tipo = Models.Enums.TipoMovimento.Receita;
-            valorProfissional.ContratoId = contrato.Id;
-            valorProfissional.DtVencimento = DateTime.Parse(await DefinirDataVencimento(movimento));
-
-            await _movimentoRepository.Adicionar(movimento);
-            return true;
+            Notificar($"A mensalidade para a competência {movimento.CompetenciaMensalidade} já existe!");
+            return false;
         }
-        public async Task<bool> Quitar(Guid id)
+
+        var mensalidade = CriarMensalidade(movimento, contratoFinanceiro);
+
+        await _movimentoRepository.Adicionar(mensalidade);
+        return true;
+    }
+
+    public async Task<bool> AdicionarValorProfissional(MovimentoFinanceiroEntidade movimento)
+    {
+        var contratoFinanceiro = await ObterContrato(movimento);
+
+        if (contratoFinanceiro == null)
         {
-            var movimento = await _movimentoRepository.BuscarMovimentoPorId(id);
-
-            if (!VerificarCondicoesQuitacao(movimento)) return false;
-            else
-            {
-                if (movimento.Contrato.Entidade.Classe == Models.Enums.ClasseEntidade.Aluno)
-                {
-                    movimento.DtPagamento = DateTime.Now;
-                    movimento.ValorPago = VerificarClasseValorReceber(movimento, movimento.Contrato);
-                    movimento.ValorReceber = 0;
-                    movimento.Situacao = Models.Enums.SituacaoMovimento.Quitado;
-                    movimento.CompetenciaPagamento = DefinirCompetenciaPagamento();
-
-                    await _movimentoRepository.Atualizar(movimento);
-                    return true;
-                }
-                else if (movimento.Contrato.Entidade.Classe == Models.Enums.ClasseEntidade.Profissional)
-                {
-                    movimento.DtPagamento = DateTime.Now;                    
-                    movimento.ValorPago = VerificarClasseValorReceber(movimento, movimento.Contrato);
-                    movimento.ValorReceber = 0;
-                    movimento.Situacao = Models.Enums.SituacaoMovimento.Quitado;
-                    movimento.CompetenciaPagamento = DefinirCompetenciaPagamento();
-
-                    await _movimentoRepository.Atualizar(movimento);
-                    return true;
-                }
-                else
-                {
-                    Notificar("Classe não permitida no processo de quitação");
-                    return false;
-                }                
-            }
-
+            Notificar("Contrato financeiro não encontrado");
+            return false;
         }
-        public async Task<bool> desquitar(Guid id)
+
+        if (EntidadeInativa(contratoFinanceiro.Entidade.DataSaida))
         {
-            var movimento = await _movimentoRepository.BuscarMovimentoPorId(id);            
-
-            if (!VerificarCondicoesDesquitacao(movimento)) return false;
-            else
-            {
-                if (movimento.Contrato.Entidade.Classe == Models.Enums.ClasseEntidade.Aluno) 
-                {
-                    movimento.DtPagamento = DateTime.Parse("0001-01-01");
-                    movimento.ValorReceber = VerificarClasseValorReceber(movimento, movimento.Contrato);
-                    movimento.ValorPago = 0;
-                    movimento.Situacao = Models.Enums.SituacaoMovimento.Aberto;
-                    movimento.CompetenciaPagamento = null;
-
-                    await _movimentoRepository.Atualizar(movimento);
-                    return true;
-                }
-                else if (movimento.Contrato.Entidade.Classe == Models.Enums.ClasseEntidade.Profissional)
-                {
-                    movimento.DtPagamento = DateTime.Parse("0001-01-01");
-                    movimento.ValorPago = 0;
-                    movimento.ValorReceber = VerificarClasseValorReceber(movimento, movimento.Contrato);
-                    movimento.Situacao = Models.Enums.SituacaoMovimento.Aberto;
-                    movimento.CompetenciaPagamento = null;
-
-                    await _movimentoRepository.Atualizar(movimento);
-                    return true;
-                }
-                else
-                {
-                    Notificar("Classe não permitida no processo de desquitação");
-                    return false;
-                }
-            }
-
+            Notificar("Entidade Inativa");
+            return false;
         }
-        public async Task<bool> Remover(Guid Id)
+
+        if (!await ExisteCompetencia(movimento))
         {
-            var movimento = await _movimentoRepository.ObterPorId(Id);
-            if (VerificarCondicoesExclusao(movimento))
-            {
-                await _movimentoRepository.Remover(Id);
-                return true;
-            }
-            else return false;
+            Notificar("Esse Profissional já possui valor para a competência " + movimento.CompetenciaMensalidade);
+            return false;
         }
-        private string DefinirCompetenciaPagamento()
+
+        var profissional = CriarValorProfissional(movimento, contratoFinanceiro);
+
+        await _movimentoRepository.Adicionar(movimento);
+        return true;
+    }
+
+    public async Task<bool> QuitarMensalidade(MovimentoFinanceiroEntidade movimento)
+    {
+        if (movimento == null)
         {
-            return (DateTime.Now.Month.ToString() + DateTime.Now.Year.ToString()).PadLeft(6, '0');
+            Notificar("Movimento inválido");
+            return false;
         }
-        private async Task<bool> VerificarCompetencia(MovimentoFinanceiroEntidade movimento)
+
+        if (!VerificarCondicoesQuitacao(movimento))
         {
-            if (await _movimentoRepository.BuscarMovimentosPorCompetencia(movimento) != null)
-                return false;
+            return false;
+        }
+
+        DefinirQuitacao(movimento);
+
+        await _movimentoRepository.Atualizar(movimento);
+        return true;
+    }
+
+    public async Task<bool> Desquitar(Guid id)
+    {
+        var movimento = await _movimentoRepository.BuscarMovimentoPorId(id);
+
+        if (movimento == null)
+        {
+            Notificar("Movimento não encontrado!");
+            return false;
+        }
+
+        if (!VerificarCondicoesDesquitacao(movimento))
+        {
+            Notificar("Não é possível desquitar esse movimento!");
+            return false;
+        }
+
+        Reabrir(movimento);
+
+        await _movimentoRepository.Atualizar(movimento);
+        return true;
+    }
+
+    public async Task<bool> Remover(Guid Id)
+    {
+        var movimento = await _movimentoRepository.ObterPorId(Id);
+
+        if (!VerificarCondicoesExclusao(movimento))
+        {
+            Notificar("Não é possível excluir esse movimento!");
+            return false;
+        }
+
+        await _movimentoRepository.Remover(Id);
+        return true;
+    }
+
+    private string DefinirCompetenciaPagamento()
+    {
+        return DateTime.Now.ToString("yyyyMM");
+    }
+
+    private async Task<bool> ExisteCompetencia(MovimentoFinanceiroEntidade movimento)
+    {
+        if (await _movimentoRepository.BuscarMovimentosPorCompetencia(movimento) != null)
+        {
+            Notificar("Movimento para essa competência já existe!");
+            return false;
+        }
+
+        return true;
+    }
+
+    private void Reabrir(MovimentoFinanceiroEntidade movimento)
+    {
+        movimento.DataPagamento = DateTime.MinValue;
+        movimento.ValorReceber = VerificarClasseValorReceber(movimento, movimento.Contrato);
+        movimento.ValorPago = default;
+        movimento.Situacao = SituacaoMovimento.Aberto;
+        movimento.CompetenciaPagamento = string.Empty;
+    }
+
+    private string DefinirDataVencimento(MovimentoFinanceiroEntidade movimento, ContratoFinanceiro contratoFinanceiro)
+    {
+        return contratoFinanceiro.Vencimento + "/" + movimento.CompetenciaMensalidade.Substring(0, 2) + "/" + DateTime.Now.Year;
+    }
+
+    private async Task<ContratoFinanceiro> ObterContrato(MovimentoFinanceiroEntidade movimento)
+    {
+        return await _contratoRepository.ObterContratoPorId(movimento.ContratoId);
+    }
+
+    private bool EntidadeInativa(DateTime dataSaida)
+    {
+        if (dataSaida != DateTime.MinValue)
+        {
             return true;
         }
-        private async Task<string> DefinirDataVencimento(MovimentoFinanceiroEntidade movimento)
-        {
-            var contrato = await ObterContrato(movimento);
-            return contrato.Vencimento + "/" + movimento.CompetenciaMensalidade.Substring(0, 2) + "/" + DateTime.Now.Year;
-        }
-        private async Task<ContratoFinanceiro> ObterContrato(MovimentoFinanceiroEntidade movimento)
-        {
-            return await _contratoRepository.ObterContratoPorId(movimento.ContratoId);
-        }
-        public void Dispose()
-        {
-            _movimentoRepository?.Dispose();
-            _contratoRepository?.Dispose();
-        }
 
-        
+        return false;
+    }
+
+    private MovimentoFinanceiroEntidade DefinirQuitacao(MovimentoFinanceiroEntidade movimento)
+    {
+        movimento.DataPagamento = DateTime.Now;
+        movimento.ValorPago = VerificarClasseValorReceber(movimento, movimento.Contrato);
+        movimento.ValorReceber = 0;
+        movimento.Situacao = SituacaoMovimento.Quitado;
+        movimento.CompetenciaPagamento = DefinirCompetenciaPagamento();
+
+        return movimento;
+    }
+
+    private MovimentoFinanceiroEntidade CriarValorProfissional(MovimentoFinanceiroEntidade movimento, ContratoFinanceiro contratoFinanceiro)
+    {
+        var valorReceber = ((contratoFinanceiro.ValorUnitario * contratoFinanceiro.Quantidade) * (contratoFinanceiro.MargemLucro) / 100);
+        var valorMensal = contratoFinanceiro.Quantidade * contratoFinanceiro.ValorUnitario;
+
+        var dataVencimento = DateTime.Parse(DefinirDataVencimento(movimento, contratoFinanceiro));
+
+        return new MovimentoFinanceiroEntidade(
+            ClasseMovimento.Profissional,
+            valorReceber,
+            valorPago: 0,
+            valorMensal,
+            0,
+            SituacaoMovimento.Aberto,
+            dataVencimento,
+            DateTime.Now,
+            contratoFinanceiro.Id,
+            TipoMovimento.Receita,
+            competenciaPagamento: movimento.CompetenciaPagamento,
+            quantidadeAlunos: contratoFinanceiro.Quantidade);
+    }
+
+    private MovimentoFinanceiroEntidade CriarMensalidade(MovimentoFinanceiroEntidade movimento, ContratoFinanceiro contratoFinanceiro)
+    {
+        var desconto = contratoFinanceiro.ValorMensal * contratoFinanceiro.Desconto / 100;
+        var valorReceber = contratoFinanceiro.ValorMensal - desconto;
+
+        var dataVencimento = DateTime.Parse(DefinirDataVencimento(movimento, contratoFinanceiro));
+
+        return new MovimentoFinanceiroEntidade(
+            ClasseMovimento.Aluno,
+            valorReceber,
+            valorPago: 0,
+            contratoFinanceiro.ValorMensal,
+            desconto,
+            SituacaoMovimento.Aberto,
+            dataVencimento,
+            DateTime.Now,
+            contratoFinanceiro.Id,
+            TipoMovimento.Receita,
+            competenciaMensalidade: movimento.CompetenciaMensalidade
+        );
+    }
+
+    public void Dispose()
+    {
+        _movimentoRepository?.Dispose();
+        _contratoRepository?.Dispose();
     }
 }
